fix: config fixes
